Use PyPy to calculate outliers. This splits the plot_outliers_by_threshold script into two: one script to generate outlier data, and one script to plot that data. Fixes #151.
